requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.4.0
pydantic==2.7.1
pydantic-settings==2.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-dotenv==1.0.1
paho-mqtt==2.1.0
redis==5.0.7
pymongo==4.7.3
